chore: 회원 정보 조회 엔드포인트 및 메서드 네이밍 수정
