+++ src/modules/auth/auth.module.ts
@@ -3,7 +3,7 @@ import { ConfigService } from '@nestjs/config';
 import { JwtModule } from '@nestjs/jwt';
 import { AuthController } from 'src/presentation/controller/auth/auth.controller';
 import { AuthService } from 'src/domain/services/auth/auth.service';
-import { UserModule } from 'src/modules/user/user.module';
+import { UserModule } from 'src/modules/users/users.module';
 import { OauthContext } from 'src/infrastructure/auth/context/auth-context';
 import { GoogleStrategy } from 'src/infrastructure/auth/strategy/google.strategy';
