+++ src/modules/users/users.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { UserController } from 'src/presentation/controller/user/user.controller';
-import { UserService } from 'src/domain/services/user/user.service';
+import { UserController } from 'src/presentation/controller/users/users.controller';
+import { UserService } from 'src/domain/services/users/users.service';
 import { UserRepository } from 'src/domain/interface/user.repository';
 import { UserPrismaRepository } from 'src/infrastructure/repositories/user-prisam.repository';
