--- src/presentation/dto/user/request/search-user.request.ts
@@ -1,3 +1,0 @@
-export class SearchUserRequest {
-    readonly searchUserId: string;
-}
