+++ src/presentation/dto/users/response/get-me.response.ts
@@ -1,6 +1,6 @@
 import { Provider } from 'src/shared/types';
 
-export class SearchMyProfileResponse {
+export class GetMeResponse {
     readonly id: string;
 
     readonly email: string;
